so, the XML.java file is more or less finished. now the constructor accepts a File instead of nothing. the xml Skeleton goes something like this:
<Sources>
	<Source value="source_value">
		<Transactions>
			<Transaction datetime="time">
				<description>description</description>
				<amount>
					<currency type="coinType"
					int="amount" />
				</amount>
			</Transaction>
		</Transactions>
		<Sum>
			<currency type="coinType" int="amount" />
		</Sum>
	</Source>
</Sources>

the XML should be transferred as Transactions. transactions come from
different Sources. like my wallet, or my website or whatever. I've added
an optional functionality for multiple coin Types, since I am going to
be using many coin Types and decentralized currencies. I also plan on
making my own digital currency. so, each transaction contains the
following information :

it has the attribute datetime, which is the time the transaction was
conducted in miliseconds.
it's child elements are description, which describes what was the
transaction, and amount. amount can contain multiple currencies. amount
has two attributes, type and int. int is an int, describing the amount.
later I might change it to be a more general float, in which case it
won't be an attribute but text content. that seems more reasonable
instead of having it as an attribute. so, anyways, int is the actual
amount,
coinType is the name of the currency.

to store and transfer the Transactions, I've created a Transaction
class. it's data contains the datetime, description, and amount.
amount is an array of Currencies. Currency is a nested class inside
Transaction which contains a coinType and an Integer. Thought about
using a map, but decided it would be wasteful.

so, that's the xml. I've decided a better Idea would be to use a
database fit for an extendible real time application, and XML just from
transferring the data. I'll be using mySQL since it's open source. asides from that I'll create another form
of mediatory data to interact with the AI. I need to be able to quickly
access the information so It'll be organized based on time and other
things, so it's a good idea to make copies of the same information. I hope mySQL
contains algorithms for log(n) search on organized tables so I won't need
to do it. anyways, it's open source. so, as always, I'll try editing it
(probably with little success, but try nonetheless).

after all of that, I'll continure reading the source of javaFX. there's
a library for 3d graphics which I'll be using, and I'll have to read and
understand that one as well. then I'll nest it inside a python
controller as an ongoing subprocess and start potentially working on the
AI and the server. I have a custom server, but I think I'll use an
apache server, since it's probably the same only with more code, more
effecient code, and less bugs ~uMu~ (also, it's open source lol).
I'll see about using Pandas and numpy effeciently for scientific
computation. I have read a few books about AI. I need to organize the
information and make notes. I want to learn more mathematics and do
excercises from books used in UNI. maybe actually sign up for a
university. still need money though... ~UmU~